+++ src/main/java/superhero/dao/LocationDaoDb.java
@@ -2,7 +2,15 @@ package superhero.dao;
 
 import superhero.model.Location;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 public class LocationDaoDb implements LocationDao {
     @Override
@@ -28,5 +36,22 @@ public class LocationDaoDb implements LocationDao {
     @Override
     public void deleteLocationById(int locationId) {
 
+    }
+
+    public static final class LocationMapper implements RowMapper<Location>{
+        @Override
+        public Location mapRow (ResultSet rs, int index) throws SQLException{
+            Location location = new Location();
+            location.setLocationId(rs.getInt("locationId"));
+            location.setLocationName(rs.getString("locationName"));
+            location.setLocationDescription(rs.getString("locationDescription"));
+            location.setLocationAddress(rs.getString("locationAddress"));
+            location.setLocationState(rs.getString("locationState"));
+            location.setLocationCity(rs.getString("locationCity"));
+            location.setLocationZip(rs.getInt("locationZip"));
+            location.setLocationLat(rs.getDouble("locationLatLat"));
+            location.setLocationLong(rs.getDouble("locationLong"));
+            return location;
+        }
     }
 }
