+++ src/main/java/superhero/dao/SuperOrganizationDaoDb.java
@@ -1,8 +1,11 @@
 package superhero.dao;
 
+import org.springframework.jdbc.core.RowMapper;
 import superhero.model.Super;
 import superhero.model.SuperOrganization;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.List;
 
 public class SuperOrganizationDaoDb implements SuperOrganizationDao {
@@ -34,5 +37,16 @@ public class SuperOrganizationDaoDb implements SuperOrganizationDao {
     @Override
     public List<Super> getSuperOrganizationForSuper(Super superHero) {
         return null;
+    }
+
+    public static final class SuperOrganizationMapper implements RowMapper<SuperOrganization> {
+        @Override
+        public SuperOrganization mapRow (ResultSet rs, int index) throws SQLException {
+            SuperOrganization superOrganization = new SuperOrganization();
+            superOrganization.setOrganizationId(rs.getInt("organizationId"));
+            superOrganization.setOrganizationName(rs.getString("organizationName"));
+            superOrganization.setOrganizationDescription(rs.getString("organizationDescription"));
+            return superOrganization;
+        }
     }
 }
